feat(instance) Move `instance.rs` into `instance/`.
